+++ CHANGELOG.rst
@@ -5,10 +5,22 @@ CHANGELOG
 0.3
 ===
 
+* New utility methods supported by Zipper
+
+  * ``drop``
+  * ``slice``
+  * ``splitAt``
+  * ``take``
+  
 * Deep-select now returns a ``Group`` rather than a ``Zipper`` with an invalid
   context
+* Implicit conversions from ``String`` and ``Symbol`` to ``Selector`` are now
+  hidden in the ``Selector`` companion object
+* Explicit converters now use ``convert`` instead of ``anti``
 * **Bug Fixes**
 
+  * `Issue #19`_ – Attributes Should Be Order-Preserving
+  * `Issue #26`_ – Zipper unselect is incorrect after multiple operations
   * ``Zipper#unselect`` fails to rebuild siblings at the second level when some
     results have been dropped
   * ``Zipper#unselect`` on empty selection results no functions appropriately
@@ -19,6 +31,8 @@ CHANGELOG
 	  by splitting it entirely from ``CanBuildFrom``.
 
 
+.. _Issue #19: https://github.com/djspiewak/anti-xml/issues/19
+.. _Issue #26: https://github.com/djspiewak/anti-xml/issues/26
 .. _Issue #12: https://github.com/djspiewak/anti-xml/issues/12